services/ia-agent/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ia-agent"
version = "0.1.0"
description = "Agente IA con RAG sobre Qdrant y proveedor LLM (OpenAI/Gemini)"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "pydantic>=2.9",
    "pydantic-settings>=2.6",
    "openai>=1.54",
    "google-generativeai>=0.8",
    "qdrant-client>=1.12",
    "aio-pika>=9.4",
    "tiktoken>=0.8",
    "structlog>=24.4",
    "httpx>=0.27",
    "pyjwt[crypto]>=2.9",
    "hotelstaff-shared",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3",
    "pytest-asyncio>=0.24",
    "pytest-cov>=5.0",
    "httpx>=0.27",
    "ruff>=0.7",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "B", "UP", "SIM", "RUF"]
ignore = ["B008"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
addopts = "-q --strict-markers"
testpaths = ["tests"]

[tool.coverage.run]
source = ["app"]
branch = true

[tool.coverage.report]
show_missing = true
fail_under = 70
